Move to size_t and other unsigned types, over off_t
We never expect the offset to be negative, so instead of keeping
it around as off_t,  and having to cast to size_t to deal with
signed vs unsigned comparisons, we just use size_t directly.

There's still a few uses of off_t and ssize_t around, but
those will need further refactoring to clean up.